+++ src/Symfony/Component/Filesystem/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * support for passing a `null` value to `Filesystem::isAbsolutePath()` is deprecated and will be removed in 5.0
+ * `Filesystem::joinPaths()` added
 
 4.3.0
 -----